REFACTOR: extract base test case
